Refactored names to keep everything organized and readable. Also, this has Firebase information implemented into the Gradle Build.
