+++ app/src/main/java/com/example/gumdropv01/ui/my_list/MyListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.gumdropv01.ui.tools
+package com.example.gumdropv01.ui.my_list
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,20 +10,20 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.example.gumdropv01.R
 
-class ToolsFragment : Fragment() {
+class MyListFragment : Fragment() {
 
-    private lateinit var toolsViewModel: ToolsViewModel
+    private lateinit var myListViewModel: MyListViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        toolsViewModel =
-            ViewModelProviders.of(this).get(ToolsViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_tools, container, false)
-        val textView: TextView = root.findViewById(R.id.text_tools)
-        toolsViewModel.text.observe(this, Observer {
+        myListViewModel =
+            ViewModelProviders.of(this).get(MyListViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
+        val textView: TextView = root.findViewById(R.id.text_gallery)
+        myListViewModel.text.observe(this, Observer {
             textView.text = it
         })
         return root
