+++ app/src/main/java/com/example/gumdropv01/ui/my_list/MyListViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.gumdropv01.ui.gallery
+package com.example.gumdropv01.ui.my_list
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class MyListViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is gallery Fragment"
