+++ app/src/main/java/com/example/gumdropv01/ui/santa_list/SantaListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.gumdropv01.ui.gallery
+package com.example.gumdropv01.ui.santa_list
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,20 +10,20 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.example.gumdropv01.R
 
-class GalleryFragment : Fragment() {
+class SantaListFragment : Fragment() {
 
-    private lateinit var galleryViewModel: GalleryViewModel
+    private lateinit var santaListViewModel: SantaListViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        galleryViewModel =
-            ViewModelProviders.of(this).get(GalleryViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
-        val textView: TextView = root.findViewById(R.id.text_gallery)
-        galleryViewModel.text.observe(this, Observer {
+        santaListViewModel =
+            ViewModelProviders.of(this).get(SantaListViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_slideshow, container, false)
+        val textView: TextView = root.findViewById(R.id.text_slideshow)
+        santaListViewModel.text.observe(this, Observer {
             textView.text = it
         })
         return root
