+++ app/src/main/java/com/example/gumdropv01/ui/santa_list/SantaListViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.gumdropv01.ui.tools
+package com.example.gumdropv01.ui.santa_list
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ToolsViewModel : ViewModel() {
+class SantaListViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is tools Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
