+++ app/src/main/java/com/example/gumdropv01/ui/settings/SettingsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.gumdropv01.ui.slideshow
+package com.example.gumdropv01.ui.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class SettingsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is tools Fragment"
     }
     val text: LiveData<String> = _text
 }
