+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,19 +13,19 @@
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.example.gumdropv01.ui.gallery.GalleryFragment"
+        android:name="com.example.gumdropv01.ui.my_list.MyListFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.example.gumdropv01.ui.slideshow.SlideshowFragment"
+        android:name="com.example.gumdropv01.ui.santa_list.SantaListFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 
     <fragment
         android:id="@+id/nav_tools"
-        android:name="com.example.gumdropv01.ui.tools.ToolsFragment"
+        android:name="com.example.gumdropv01.ui.settings.SettingsFragment"
         android:label="@string/menu_tools"
         tools:layout="@layout/fragment_tools" />
